resume-ranker/backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
supabase
python-dotenv
